% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgkb.R
\name{pgkb_context}
\alias{pgkb_context}
\title{Generate LLM-friendly context for the api.pharmgkb.org package}
\usage{
pgkb_context()
}
\value{
Character string (invisibly), also printed
}
\description{
Generate LLM-friendly context for the api.pharmgkb.org package
}
