% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgkb.R
\name{pgkb_variant}
\alias{pgkb_variant}
\title{Look up a genetic variant in PharmGKB}
\usage{
pgkb_variant(rsid)
}
\arguments{
\item{rsid}{Variant rsID (e.g. "rs1065852", "rs4149056", "rs9923231")}
}
\value{
tibble: id, symbol, name, change_classification,
clinical_significance, rare, type
}
\description{
Look up a genetic variant in PharmGKB
}
